Rename CollectionProxy to ProxyMixin and use this name consistently
Credit to @ar-jan in #1 for pointing out the former inconsistency.
