+++ index.js
@@ -1,3 +1,3 @@
-export { default as CollectionProxy } from './src/collection-proxy.js';
+export { default as ProxyMixin } from './src/proxy-mixin.js';
 export { default as deriveFiltered } from './src/filtered-collection.js';
 export { default as deriveMapped } from './src/mapped-collection.js';
